requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
langchain==0.2.14
langchain-openai==0.1.15
langchain-community==0.2.10
langchain-core==0.2.35
langchain-postgres>=0.0.9
pgvector>=0.2.5
redis>=5.0.0
tiktoken>=0.7.0
pydantic-settings>=2.3.0
